refactor: extract usePoiMedia hook + role-based sidebar rendering (#184)
Move the sidebar's duplicated media state/effects/refresh into a shared
usePoiMedia hook. The hook now refreshes the POI object for every POI
type, not just destinations — fixing the linear-feature media bug noted
in issue #184 (AC: "media handling is identical across all POI types").

Convert sidebar display reads from feature_type to poi_roles.includes()
per NFR-019-02 (edit-form feature_type fields, which edit a real column,
are intentionally left).

Phase C (part 2) of the frontend unification (spec 019).
